Migrated from Portage:  - TM and LM filtering enhancements a la Badr et al 2007.  - many optimizations, including openmp-based parallelism, in rescoring.  - many more new features in rescoring.

+++ src/canoe/phrase_tm_align.cc
@@ -220,7 +220,7 @@ int main(Uint argc, const char * const * argv)
          split(line, src_sent, " ");
       }
       else {
-         if (i == (int)src_sents.size())
+         if (i == src_sents.size())
             break;
          src_sent = src_sents[i];
       }
